Run futurize -2 -x division_safe -w .

+++ python/lsst/meas/base/forcedPhotCoadd.py
@@ -1,3 +1,4 @@
+from builtins import zip
 #!/usr/bin/env python
 #
 # LSST Data Management System
@@ -105,11 +106,11 @@ class ForcedPhotCoaddTask(ProcessImageForcedTask):
         # allow us to revert back to the old behavior of generating new forced source IDs,
         # just by renaming the ID in config.copyColumns to "object_id".
         expBits = dataRef.get(self.config.coaddName + "CoaddId_bits")
-        expId = long(dataRef.get(self.config.coaddName + "CoaddId"))
+        expId = int(dataRef.get(self.config.coaddName + "CoaddId"))
         return lsst.afw.table.IdFactory.makeSource(expId, 64 - expBits)
 
     def getExposureId(self, dataRef):
-        return long(dataRef.get(self.config.coaddName + "CoaddId"))
+        return int(dataRef.get(self.config.coaddName + "CoaddId"))
 
     def fetchReferences(self, dataRef, exposure):
         """Return an iterable of reference sources which overlap the exposure
